Nest the WebDriver satellites under their concept
#1770 nested satellite types under the concept they belong to across the
repository — `XmlError` → `Xml.Error`, `TypescriptError` →
`Typescript.Error` — and gave tarantula `WebDriverError` →
`WebDriver.Error`. This branch had reintroduced the flat name, so this
restores it and carries the convention through the three satellites the
rewrite added: `WebDriver.Session`, `WebDriver.Element` and
`WebDriver.ShadowRoot`, all merged into `tarantula.WebDriver.scala` as
`xenophile.Typescript.scala` was.

`soundness` now gains one type name from this module where it briefly
gained six. It also retires the `WebElement` spelling, which existed
only to dodge honeycomb's `Element`: nesting makes that collision go
away on its own terms.

Three things to know if you touch this file:

  - There is no file-level `import WebDriver.Error.Reason`. It forces
    `object WebDriver` to complete while its own members are still being
    typed, and cycles through `given sessional`, whose type mentions
    `Session`. `Reason` is written `Error.Reason` inside the object.
  - For the same reason `case class Error` names its parameter type as
    `Error.Reason`, not `WebDriver.Error.Reason`. The fully qualified
    form cycles; the sibling reference does not.
  - `class WebDriver` stays at top level as the object's companion; only
    the satellites are nested.
