+++ doc/errors/589.md
@@ -1,4 +1,4 @@
-# `SN-589` — `WebDriverError`
+# `SN-589` — `WebDriver.Error`
 
 Raised when Tarantula's WebDriver action fails inside the browser.
 
@@ -59,4 +59,4 @@ since append from 9.
 
 ## See also
 
-- Source: `lib/tarantula/src/core/tarantula.WebDriverError.scala`
+- Source: `lib/tarantula/src/core/tarantula.WebDriver.scala`
