--- lib/tarantula/src/core/tarantula.WebElement.scala
@@ -1,60 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package tarantula
-
-import anticipation.*
-import jacinta.*
-import prepositional.*
-
-object WebElement:
-  // Elements cross the wire as `{"element-6066-…": id}` — as a script argument, or as a
-  // pointer action's origin. `ElementRef` carries the key, not a legal Scala identifier.
-  given encodable: (WebElement is Encodable in Json) =
-    element => ElementRef(element.elementId).in[Json]
-
-// An element handle: the opaque reference a driver returned for a node, meaningful only to the
-// session that produced it.
-//
-// A *pure* value, holding neither a capability nor a reference to its session — a deliberate
-// choice, not an omission. `session.elements(selector)` builds elements inside a `map` lambda,
-// and a capability-classed `WebElement` constructed there is fresh at each construction, so a
-// `List` of them cannot be formed ("any² is not visible from any"); that is precisely the
-// failure recorded in rep/DECISIONS.md which caused the earlier attempt to be reverted. The
-// `X^{this, caps.any}` borrow that `HttpSession#fetch` uses works for one value returned from a
-// method; it does not survive being built in a lambda and collected into a container.
-//
-// Nothing is lost by leaving elements pure. A leaked element is a dangling id, not a live
-// handle, and every operation on one needs a session in contextual scope — which capture
-// checking confines to its block. Using a stale one now raises `Reason.StaleElement`, where
-// before it was silently swallowed.
-case class WebElement(elementId: Text)
